issue #76: parse rain values as float instead integer

+++ main.js
@@ -1726,7 +1726,7 @@ async function insertIntoList(key, value, unit, newObj = null) {
                         }
                     };
                 } else if (key.match(/\.rain_value/) || key.match(/\.rain/)) {
-                    valueType = "int";
+                    valueType = "float";
                     obj = {
                         type: "state",
                         common: {
